requirements.txt
flask
uwsgi
werkzeug
flask_jwt_extended
flask_sqlalchemy
pycrypto
psycopg2
flask_socketio
ujson
datetime